+++ CHANGELOG.md
@@ -10,7 +10,7 @@
   * Add `PayPalNativeShippingListener` to receive events on changes in shipping information. Add `PayPalNativeShippingAddress`, `PayPalNativeShippingMethod` and `PayPalNativePaysheetActions`
   * Remove `PayPalCheckout` as an `api` dependency
 * CorePayments:
-  * Send `orderID` instead of `sessionID` for analytics
+  * Send `orderId` instead of `sessionId` for analytics
 
 ## 0.0.8 (2023-03-13)
 * `PayPalNativePayments`: